doc97: Decrypt the `Data' stream as well.
According to the official documentation [1], the `Data' stream must be
encrypted and thus, decrypted as well. I tried with a doc file with
lots of pictures, when the `Data' stream is not decrypted, complaint
about corrupted data is made, and none of the pictures could be
viewed, on the other hand everything works fine if it's fully decrypted.

[1] https://msdn.microsoft.com/en-us/library/dd945648(v=office.12).aspx
    https://msdn.microsoft.com/en-us/library/dd946845(v=office.12).aspx